Update build_logistic_regression_classifier.ipynb
Script now reports results of training the model on all the datasets. Often data is held in reserve for validating a model. For this reason, the data shall be processed once again, but using all the input data as training data. The results can then be used to confirm the run_logistic_regression_classifier.ipynb
is using coefficients correctly to make predictions.